# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BACKEND CART API",
        "description": "Morsum challenge",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/carts": {
            "post": {
                "tags": [
                    "carts"
                ],
                "description": "Registers a new cart",
                "operationId": "e2e275c2483722396fe380b18fcc77e7",
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "success": {
                                        "summary": "Provides id of the created cart",
                                        "value": {
                                            "uuid": "71cca617-e92e-46ae-9394-74582ec7d00a"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Unauthenticated user",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/carts/{uuid}": {
            "get": {
                "tags": [
                    "carts"
                ],
                "description": "Get items of cart",
                "operationId": "634a0f6045dc19767f89b2e95fa035ad",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Cart universally unique identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of cart items",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "success": {
                                        "summary": "List of cart items",
                                        "value": {
                                            "data": [
                                                {
                                                    "items": [
                                                        {
                                                            "name": "Product X",
                                                            "sku": "PDTOMUTXPPZ",
                                                            "description": "Some description here.",
                                                            "price": 100,
                                                            "quantity": 1
                                                        }
                                                    ],
                                                    "total_items": 1,
                                                    "total_amount": 100
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Unauthenticated user",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Resource not found",
                                        "value": {
                                            "message": "Cart not found"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "carts"
                ],
                "description": "Add item to cart",
                "operationId": "a506e7d1296d710ae60dc0bc9edcabb0",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "Cart universally unique identifier",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "sku": {
                                        "type": "string"
                                    },
                                    "quantity": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "success": {
                                        "summary": "Provides id of the created cart",
                                        "value": {
                                            "uuid": "71cca617-e92e-46ae-9394-74582ec7d00a"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Unauthenticated user",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Resource not found",
                                        "value": {
                                            "message": "Cart not found"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Form validation errors"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/products": {
            "get": {
                "tags": [
                    "products"
                ],
                "description": "Get product list",
                "operationId": "072fd15d3be55577bb08c19d19f1e75c",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Maximum number of results to return per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns a list of products",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "success": {
                                        "summary": "List of products",
                                        "value": {
                                            "data": [
                                                {
                                                    "id": 1,
                                                    "name": "Product X",
                                                    "sku": "PDTOMUTXPPZ",
                                                    "description": "Some description here",
                                                    "price": 123.4500000000000028421709430404007434844970703125
                                                }
                                            ]
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid token",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "error": {
                                        "summary": "Unauthenticated user",
                                        "value": {
                                            "message": "Unauthenticated."
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Form validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/register": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "Registers a new user",
                "operationId": "682fc9d9ee8343ea3f2fa34b923f8722",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation"
                    },
                    "422": {
                        "description": "Form validation errors"
                    }
                }
            }
        },
        "/api/v1/tokens": {
            "post": {
                "tags": [
                    "users"
                ],
                "description": "An registered user can issue a token to make API calls",
                "operationId": "7b08757f05c65b8778c1469537d006f1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string"
                                    },
                                    "password": {
                                        "type": "string"
                                    },
                                    "device_name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "examples": {
                                    "success": {
                                        "summary": "A valid token is issued to the authenticated user",
                                        "value": {
                                            "access_token": "valid-token-here"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "The credentials providaded are incorrect or does not exist on the database"
                    },
                    "422": {
                        "description": "Form validation errors"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}